Fix backup bug
zipFile.Open() no longer throws an exception in this configuration, but the fileInfo object stores information whether the file we are trying to backup exists.
Reminder: This method does a backup of the target directories, only if there are some missing files. If we are overwriting a previous installation, there is no real reason to do a backup and waste space.
Caution: May be a cause of a bug, where valuable data doesn't get backed up!